+++ examples/resources/pingfederate_authentication_policies_settings/import.sh
@@ -1,2 +1,2 @@
-# "authenticationPoliciesSettingsId" should be "id"
+# This resource is singleton, so the value of "id" doesn't matter - it is just a placeholder, and required by Terraform
 terraform import pingfederate_authentication_policies_settings.myAuthenticationPoliciesSettings authenticationPoliciesSettingsId