+++ examples/resources/pingfederate_idp_default_urls/import.sh
@@ -1,2 +1,2 @@
-# "idpDefaultUrlId" should be "id"
+# This resource is singleton, so the value of "id" doesn't matter - it is just a placeholder, and required by Terraform
 terraform import pingfederate_idp_default_urls.myIdpDefaultUrl idpDefaultUrlId