+++ examples/resources/pingfederate_license_agreement/import.sh
@@ -1,2 +1,2 @@
-# "licenseAgreementId" should be "id"
+# This resource is singleton, so the value of "id" doesn't matter - it is just a placeholder, and required by Terraform
 terraform import pingfederate_license_agreement.myLicenseAgreement licenseAgreementId