+++ examples/resources/pingfederate_session_authentication_session_policies_global/import.sh
@@ -1,2 +1,2 @@
-# "sessionAuthenticationSessionPoliciesGlobalId" should be "id"
+# This resource is singleton, so the value of "id" doesn't matter - it is just a placeholder, and required by Terraform
 terraform import pingfederate_session_authentication_session_policies_global.mySessionAuthenticationSessionPoliciesGlobal sessionAuthenticationSessionPoliciesGlobalId