+++ examples/resources/pingfederate_session_settings/import.sh
@@ -1,2 +1,2 @@
-# "sessionSettingsId" should be "id"
+# This resource is singleton, so the value of "id" doesn't matter - it is just a placeholder, and required by Terraform
 terraform import pingfederate_session_settings.mySessionSettings sessionSettingsId