+++ internal/resource/config/datastore/custom_data_store.go
@@ -57,7 +57,7 @@ func toSchemaCustomDataStore() schema.SingleNestedAttribute {
 		"plugin_descriptor_ref": schema.SingleNestedAttribute{
 			Required:    true,
 			Description: "Reference to the plugin descriptor for this instance. The plugin descriptor cannot be modified once the instance is created. Note: Ignored when specifying a connection's adapter override.",
-			Attributes:  resourcelink.ToSchema(),
+			Attributes:  resourcelink.ToSchemaLocationUseStateForUnknown(),
 		},
 		"parent_ref": schema.SingleNestedAttribute{
 			Computed:    true,